requirements.txt
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.7.0
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fasteners==0.14.1
future==0.18.2
h11==0.13.0
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.9.3
pexpect==4.8.0
pillow==10.3.0
protobuf==3.12.4
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.1
requests
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
urwid==2.1.2
uvicorn==0.15.0
wadllib==1.3.6
wsproto==1.0.0
xdg==5.0.0
zipp==3.18.1
